Create about.me
In this project, we explore the concept of mapping a list of data to individual widgets in the Flutter framework. The project's main goal is to illustrate how to create a list of items, each represented by a custom widget, and display them in a scrollable list. We achieve this by utilizing the powerful ListView.builder widget, which helps optimize performance by rendering only the visible items on the screen.